Update to built in game idea.
Created an overlay for the game to run on. wasd typed in a row on the keyboard will trigger the overlay and then there is an event listener to close the overlay when esc text is clicked. Wanted to get the button esc working but couldn't figure it out for tonight.